Refactoring of Nodes::Set::Set to Nodes::Set
