+++ lib/msfl_visitors/nodes.rb
@@ -23,7 +23,7 @@ require_relative 'nodes/named_value'
 require_relative 'nodes/number'
 require_relative 'nodes/partial'
 require_relative 'nodes/range_value'
-require_relative 'nodes/set/set'
+require_relative 'nodes/set'
 require_relative 'nodes/time'
 require_relative 'nodes/value'
 require_relative 'nodes/word'
