+++ lib/msfl_visitors/nodes/iterator.rb
@@ -8,10 +8,10 @@ module MSFLVisitors
       # Be extra defensive, because even after adding previous comment I still tend to make the mistake of
       # passing in an array.
       #
-      # @param set [MSFLVisitors::Nodes::Set::Set] a set node that allows its elements to be iterated over
+      # @param set [MSFLVisitors::Nodes::Set] a set node that allows its elements to be iterated over
       def initialize(set)
-        unless set.is_a? MSFLVisitors::Nodes::Set::Set
-          fail ArgumentError, "Argument to Iterator initialize must be instance of MSFLVisitors::Nodes::Set::Set"
+        unless set.is_a? MSFLVisitors::Nodes::Set
+          fail ArgumentError, "Argument to Iterator initialize must be instance of MSFLVisitors::Nodes::Set"
         end
         self.set = set
       end
