+++ spec/nodes/iterator_spec.rb
@@ -10,11 +10,11 @@ describe MSFLVisitors::Nodes::Iterator do
 
     let(:left) { MSFLVisitors::Nodes::Iterator.new left_set }
 
-    let(:left_set) { MSFLVisitors::Nodes::Set::Set.new [one, two] }
+    let(:left_set) { MSFLVisitors::Nodes::Set.new [one, two] }
 
     let(:right) { MSFLVisitors::Nodes::Iterator.new right_set }
 
-    let(:right_set) { MSFLVisitors::Nodes::Set::Set.new [one, two] }
+    let(:right_set) { MSFLVisitors::Nodes::Set.new [one, two] }
 
     subject { left == right }
 
@@ -27,7 +27,7 @@ describe MSFLVisitors::Nodes::Iterator do
 
       context "when lhs#items is not equal to rhs#items" do
 
-        let(:right_set) { MSFLVisitors::Nodes::Set::Set.new [one] }
+        let(:right_set) { MSFLVisitors::Nodes::Set.new [one] }
 
         it { is_expected.to be false }
       end
